Refactor "Advanced design concepts" to be more encapsulated
