+++ handlers/jobs/new.py
@@ -1,0 +1,6 @@
+from .base import BaseJobHandler
+
+
+class JobHandler(BaseJobHandler):
+    def get_queues(self):
+        return ['new', 'data']
