+++ handlers/jobs/suspended.py
@@ -1,0 +1,6 @@
+from .base import BaseJobHandler
+
+
+class JobHandler(BaseJobHandler):
+    def get_queues(self):
+        return ['suspended', 'assigned', 'data']
