+++ handlers/queues/suspended.py
@@ -1,0 +1,6 @@
+from .base import BaseQueueHandler
+
+
+class QueueHandler(BaseQueueHandler):
+    def build_path(self):
+        return f'/jobs_suspended/{self.job_object.id}/'
